Creates config.js, updates nodejs.yml, and updates google-sheets.
This commit creates the config.js file to create the creds object
on config using the process environment object. This commit updates
the nodejs.yml to pass the secrets on GitHub for the project as
environment variables. This essentailly includes the dotenv file
on the GitHub test runner. Finally, this commit udpates the google
sheets module to pull in the config's creds object instead of
using the client_secret.json file.