Updating src/app/google-sheets
This commit updates the main app to include 4 new functions. These
functions are getCell, getCells, saveCell, and saveCells. Get cell
returns an individual cell based on the row and column. Get rows
returns a rows of cells based on the beginning and ending row.
The save cell function saves as cell and updates it in the sheet
by passing it the cell. Finally, the save cells function does a
bulk update of cells that need to be saved based on the sheet and
the modified cells.